Don't always show the link to download the csv
Don't show the link when we are not filtering down by agency or vendor,
because this returns the entire data set. It takes minutes on my
computer to produce the csv.

The download to csv link does not show up if the user selected more
than 10 vendors or agencies.

This could be solved in the future with some caching / pre-rendering.